+++ tests/api_resources/test_accounts.py
@@ -1,0 +1,74 @@
+# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+from __future__ import annotations
+
+import os
+from typing import Any, cast
+
+import pytest
+
+from tests.utils import assert_matches_type
+from runloop_api_client import Runloop, AsyncRunloop
+from runloop_api_client.types import AccountView
+
+base_url = os.environ.get("TEST_API_BASE_URL", "http://127.0.0.1:4010")
+
+
+class TestAccounts:
+    parametrize = pytest.mark.parametrize("client", [False, True], indirect=True, ids=["loose", "strict"])
+
+    @parametrize
+    def test_method_me(self, client: Runloop) -> None:
+        account = client.accounts.me()
+        assert_matches_type(AccountView, account, path=["response"])
+
+    @parametrize
+    def test_raw_response_me(self, client: Runloop) -> None:
+        response = client.accounts.with_raw_response.me()
+
+        assert response.is_closed is True
+        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+        account = response.parse()
+        assert_matches_type(AccountView, account, path=["response"])
+
+    @parametrize
+    def test_streaming_response_me(self, client: Runloop) -> None:
+        with client.accounts.with_streaming_response.me() as response:
+            assert not response.is_closed
+            assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+
+            account = response.parse()
+            assert_matches_type(AccountView, account, path=["response"])
+
+        assert cast(Any, response.is_closed) is True
+
+
+class TestAsyncAccounts:
+    parametrize = pytest.mark.parametrize(
+        "async_client", [False, True, {"http_client": "aiohttp"}], indirect=True, ids=["loose", "strict", "aiohttp"]
+    )
+
+    @parametrize
+    async def test_method_me(self, async_client: AsyncRunloop) -> None:
+        account = await async_client.accounts.me()
+        assert_matches_type(AccountView, account, path=["response"])
+
+    @parametrize
+    async def test_raw_response_me(self, async_client: AsyncRunloop) -> None:
+        response = await async_client.accounts.with_raw_response.me()
+
+        assert response.is_closed is True
+        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+        account = await response.parse()
+        assert_matches_type(AccountView, account, path=["response"])
+
+    @parametrize
+    async def test_streaming_response_me(self, async_client: AsyncRunloop) -> None:
+        async with async_client.accounts.with_streaming_response.me() as response:
+            assert not response.is_closed
+            assert response.http_request.headers.get("X-Stainless-Lang") == "python"
+
+            account = await response.parse()
+            assert_matches_type(AccountView, account, path=["response"])
+
+        assert cast(Any, response.is_closed) is True